backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.0
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31
gunicorn>=21.0
whitenoise>=6.6
django-storages>=1.14
boto3>=1.34
dj-database-url>=2.1.0
PyMySQL>=1.1.0
PyGithub>=2.1.0
google-cloud-storage>=2.14.0
instaloader==4.15
TikTokApi==7.2.1
nest-asyncio==1.6.0
facebook-scraper==0.2.59
lxml[html_clean]>=5.2.0
google-generativeai>=0.3.0
